+++ CHANGELOG.md
@@ -3,9 +3,11 @@ All notable changes to **Device Detector** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Device Detector** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 1.5.1]
+## [Unreleased - will be 2.0.0]
 
 ### Added
+- [WP-CLI] New command to display Device Detector status: see `wp help device status` for details.
+- [WP-CLI] New command to toggle on/off main settings: see `wp help device settings` for details.
 - New Site Health "info" section about shared memory.
 
 ### Changed